# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API prueba previa junior PHP",
        "version": "1.0.0"
    },
    "paths": {
        "/api/deporte": {
            "get": {
                "tags": [
                    "Deporte"
                ],
                "summary": "Mostrar deportes",
                "operationId": "98e64d74f817c9f9391754494eaa1582",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los deportes.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Deporte"
                ],
                "summary": "Crear deporte",
                "operationId": "b95b5bb6542820d3b2ab52b8749ac31f",
                "requestBody": {
                    "request": "DeporteStoreRequest",
                    "description": "Contiene los datos del deporte a añadir",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "deporte": {
                                    "name": "nombreDeporte",
                                    "codigo": "NDEP"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deporte creado de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deporte/{id}": {
            "get": {
                "tags": [
                    "Deporte"
                ],
                "summary": "Mostrar un deporte",
                "operationId": "6bda5e7d609b88609325c73bd38a161a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del deporte a mostrar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un deporte.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Deporte"
                ],
                "summary": "Modificar deporte",
                "operationId": "e3bad1d3ae0535f97fb1ac475cad2740",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del deporte a modificar",
                        "required": true
                    }
                ],
                "requestBody": {
                    "request": "DeporteUpdateRequest",
                    "description": "Contiene los datos del deporte a modificar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "deporte": {
                                    "name": "nombreDeporte",
                                    "codigo": "NDEP"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deporte modificado de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Deporte"
                ],
                "summary": "Borrar un deporte",
                "operationId": "95791568bafc3b2d3ac3f3f69af53fd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del deporte a eliminar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deporte eliminado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}